More fixes to populate phoenix models correctly.
