+++ banzai_nres/utils/phoenix_utils.py
@@ -1,32 +1,22 @@
-from banzai_nres import settings
-import os
-from astropy.io import fits
-import boto3
-import io
-from banzai_nres.dbs import get_phoenix_model_record
+def plus_minus_to_pm(value, fmt):
+    output_str = f'{value:{fmt}}'
+    return output_str.replace('+', 'p').replace('-', 'm')
 
 
-class PhoenixModelLoader:
-    def __init__(self):
-        _wavelengths = self.open_model_fits_file(settings.PHOENIX_WAVELENGTH_FILE_LOCATION,
-                                                 settings.PHOENIX_WAVELENGTH_FILENAME)
+def pm_to_plus_minus(value):
+    return value.replace('p', '+').replace('m', '-')
 
-    def load(self, db_address, T_effective, log_g, metallicity, alpha):
-        # Load in the template
-        model_record = get_phoenix_model_record(db_address, T_effective, log_g, metallicity, alpha)
-        flux = self.open_model_fits_file(model_record.location, model_record.filename)
-        return {'wavelength': self._wavelengths, 'flux': flux}
 
-    @staticmethod
-    def open_model_fits_file(file_location, filename):
-        if 's3' in file_location:
-            s3 = boto3.client('s3')
-            buffer = io.BytesIO()
-            s3.download_fileobj(file_location.replace('s3://', ''), filename, buffer)
-            buffer.seek(0)
-            hdu = fits.open(buffer, memmap=False)
-        else:
-            hdu = fits.open(os.path.join(file_location, filename))
+def filename_to_parameters(filename):
+    _, T_effective, log_g, metallicity, alpha = filename.split('-')
+    T_effective = int(pm_to_plus_minus(T_effective))
+    log_g = float(pm_to_plus_minus(log_g))
+    metallicity = float(pm_to_plus_minus(metallicity))
+    alpha = float(pm_to_plus_minus(alpha))
+    return T_effective, log_g, metallicity, alpha
 
-        data = hdu[0].data.copy()
-        return data
+
+def parameters_to_filename(T_effective, log_g, metallicity, alpha):
+    filename =  f'phoenix-{plus_minus_to_pm(int(T_effective), "05d")}-{plus_minus_to_pm(log_g, "+0.1f")}'
+    filename += f'-{plus_minus_to_pm(metallicity, "+0.1f")}-{plus_minus_to_pm(alpha, "+0.1f")}.fits'
+    return filename
